// src/interfaces/ITransactionsResponseProps.ts
export interface ITransactionResponseProps {
    id: string; // Inclua o ID aqui
    type: string;
    amount: number;
    description: string;
    date: string;
  }



  
  // src/interfaces/ITransactionResponseProps.ts
export interface ITransactionResponseProps {
    id: string;
    type: string;
    amount: number;
    description: string;
    date: string;
  }

  
  